bot/requirements.txt
# Bot Telegram - Dame Hot Game
# Installation : python -m venv .venv && .venv\Scripts\activate (Windows) ou source .venv/bin/activate (Unix)
# Puis : pip install -r requirements.txt

python-telegram-bot>=21.0
